// src/shared/model/teamMembers.ts
export enum teamMembers {
    Noam_Choen = "Noam Choen  Age: 20",
    Tamar_Levi = "Tamar Levi   ",
    Eitan_Avrahan = "Eitan Avrahan Age: 19",
    Maya_David = "Maya David Age: 25",
    Yonah_Shalev = "Yonah Shalev ",
    Lior_Baruch = "Lior Baruch ",
    Shira_Choen = "Shira Choen ",
    Itai_Gilboa = "Itai Gilboa Age: 42",
  }